Rewrite all Livermore benchmarks to 1-based indexing matching Fortran
All 24 C and WL kernel files rewritten to use 1-based array indexing
that directly matches the Fortran originals, eliminating off-by-one
translation errors.

Changes:
- Loop bounds match Fortran: DO K=1,N → for k=1..N / while k<=N
- Array subscripts match Fortran: X(K) → x[k]
- Array sizes +1 for 1-based (Fortran DIMENSION X(1001) → x[1002])
- k04: x array increased to 1202 for OOB padding (matches Fortran behavior)
- k13/k14: use % instead of & (WL lacks bitwise AND)
- All defensive bounds guards removed from WL versions
- C uses long (64-bit) integers, double floats
- signel.h updated to long types
- CodeGen: handle large immediates (>4095) in cmpImm by loading to register

+++ benchmarks/livermore/signel.h
@@ -5,12 +5,12 @@
 #ifndef SIGNEL_H
 #define SIGNEL_H
 
-static void signel(double *v, int n) {
+static void signel(double *v, long n) {
     double fuzz  = 1.2345e-3;
     double buzz  = 1.0 + fuzz;
     double fizz  = 1.1 * fuzz;
     double scaled = 0.1;
-    for (int k = 0; k < n; k++) {
+    for (long k = 0; k < n; k++) {
         buzz = (1.0 - fuzz) * buzz + fuzz;
         fuzz = -fuzz;
         v[k] = (buzz - fizz) * scaled;
@@ -20,10 +20,10 @@ static void signel(double *v, int n) {
 /* Fill particle array P[512][4] with original Fortran values:
    DS = 1.0, DW = 0.5, P(j,k) = DS; DS = DS + DW
    In column-major (Fortran) order: j=1..4, k=1..512 */
-static void init_particles(double *p, int np) {
+static void init_particles(double *p, long np) {
     double ds = 1.0, dw = 0.5;
-    for (int j = 0; j < 4; j++) {
-        for (int k = 0; k < np; k++) {
+    for (long j = 0; j < 4; j++) {
+        for (long k = 0; k < np; k++) {
             p[k * 4 + j] = ds;  /* p[k][j] in C row-major */
             ds += dw;
         }
